test(web): isolate slow-link ordering proof
