+++ package.json
@@ -83,6 +83,7 @@
     "smoke:web-membership-refusal": "tsx implementations/web/smoke/membership-refusal.smoke.ts",
     "smoke:web-snapshot": "tsx implementations/web/smoke/snapshot-retention.smoke.ts",
     "smoke:web-bounded-aggregation": "tsx implementations/web/smoke/bounded-aggregation.smoke.ts",
+    "smoke:web-proxy-order": "tsx implementations/web/smoke/slow-link-serialization.probe.ts",
     "smoke:codex-installed": "pnpm --filter cotal-ai... build && tsx extensions/connector-codex/smoke/codex-installed.smoke.ts",
     "smoke:codex-live": "tsx extensions/connector-codex/smoke/codex-live.smoke.ts",
     "smoke:codex-tui-live": "tsx extensions/connector-codex/smoke/codex-tui-live.smoke.ts",
